requirements.txt
pillow
inflect
